requirements.txt
streamlit>=1.52.0
rdflib>=7.5.0
requests>=2.32.0
PyYAML>=6.0.